frontend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
pykrx>=1.0.36
google-genai>=0.4.0
openai>=1.3.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml_html_clean>=0.1.0
plotly>=5.17.0

